requirements.txt

langchain==0.3.26
langchain-google-genai>=0.0.5
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
sentence-transformers>=2.2.2
google-generativeai>=0.3.2
